requirements.txt
Django==4.0.4
psycopg2==2.9.3
django-cors-headers==3.13.0
djangorestframework==3.13.1
djoser==2.1.0
drf_spectacular==0.22.0
djangorestframework_simplejwt==5.2.0
